requirements.txt
# Core dependencies
openai>=1.12.0
tiktoken>=0.5.2  # For token counting with OpenAI models
anthropic>=0.18.0
google-generativeai>=0.3.0  # For Google Gemini models
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# CLI dependencies
rich>=13.7.0
typer>=0.9.0

# Database dependencies
sqlalchemy>=2.0.25
alembic>=1.13.1

# FastAPI dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0  # Pin to 4.x for passlib compatibility (5.x breaks passlib)
email-validator>=2.1.0
slowapi>=0.1.9  # Rate limiting

# Redis dependencies (P1.2 - Production cache)
redis>=5.0.0  # Redis client for distributed caching
hiredis>=2.2.3  # C parser for faster Redis protocol parsing (optional performance boost)

# Monitoring dependencies (HIGH-01 - Prometheus metrics)
prometheus-client>=0.19.0  # Prometheus metrics instrumentation

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.26.0

# Sprint 5 dependencies - Git N2 Traceability + Analytics
GitPython>=3.1.40  # Git repository integration
matplotlib>=3.8.0  # Visualization for reports (future)
plotly>=5.18.0     # Interactive charts for dashboards (future)
openpyxl>=3.1.2    # Excel export for reports (future)
reportlab>=4.0.7   # PDF generation for reports (future)
pandas>=2.1.4      # Data aggregation and analysis